requirements.txt
opencv-contrib-python
opencv-python~=4.4.0.46
numpy~=1.19.2
pillow~=8.0.1
matplotlib~=3.3.2
tqdm~=4.55.0